requirements.txt
djangorestframework>=3.12, <3.14
psycopg2-binary
django>3.1, <3.2
